+++ test/QC/Generators/Database.hs
@@ -13,20 +13,24 @@ randomStarKey = do
     aId <- arbitrary
     return $ toSqlKey aId
 
+
 randomStarSystemKey :: Gen (Key StarSystem)
 randomStarSystemKey = do
     aId <- arbitrary
     return $ toSqlKey aId
+
 
 randomStarLaneKey :: Gen (Key StarLane)
 randomStarLaneKey = do
     aId <- arbitrary
     return $ toSqlKey aId
 
+
 randomFactionKey :: Gen (Key Faction)
 randomFactionKey = do
     aId <- arbitrary
     return $ toSqlKey aId
+
 
 randomPlanetKey :: Gen (Key Planet)
 randomPlanetKey = do
@@ -38,7 +42,14 @@ randomUserKey = do
     aId <- arbitrary
     return $ toSqlKey aId
 
+
 randomDesignKey :: Gen (Key Design)
 randomDesignKey = do
+    aId <- arbitrary
+    return $ toSqlKey aId
+
+
+randomPersonKey :: Gen (Key Person)
+randomPersonKey = do
     aId <- arbitrary
     return $ toSqlKey aId