+++ test/QC/Generators/Import.hs
@@ -2,6 +2,7 @@ module QC.Generators.Import
     ( module QC.Generators.Construction
     , module QC.Generators.Database
     , module QC.Generators.News
+    , module QC.Generators.People
     , module QC.Generators.Planets
     , module QC.Generators.Reports
     , module QC.Generators.StarSystems
@@ -10,6 +11,7 @@ module QC.Generators.Import
 import QC.Generators.Construction
 import QC.Generators.Database
 import QC.Generators.News
+import QC.Generators.People
 import QC.Generators.Planets
 import QC.Generators.Reports
 import QC.Generators.StarSystems